# talde1erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/kolore": {
            "get": {
                "tags": [
                    "Koloreak"
                ],
                "description": "Kolore guztiak bueltatzen ditu.",
                "operationId": "80b12607f464d1ee3f6e3031dcbd98ff",
                "responses": {
                    "200": {
                        "description": "Kolore guztiak bueltatu ditu."
                    },
                    "404": {
                        "description": "Ez du kolorerik topatu."
                    }
                }
            },
            "put": {
                "tags": [
                    "Koloreak"
                ],
                "description": "Adierazitako informazioaren arabera kolore bat eguneratzen du datubasean.",
                "operationId": "757debd722d583f9c962e6a69d7c4ad6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Transakzioaren ID-a",
                                        "type": "integer"
                                    },
                                    "data": {
                                        "description": "Transakzioaren data",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "id_bezeroa": {
                                        "description": "Bezeroaren ID-a",
                                        "type": "integer"
                                    },
                                    "id_produktua": {
                                        "description": "Produktuaren ID-a",
                                        "type": "integer"
                                    },
                                    "kantitatea": {
                                        "description": "Produktu kantitatea",
                                        "type": "integer"
                                    },
                                    "bolumena": {
                                        "description": "Kolorearen bolumena",
                                        "type": "string"
                                    },
                                    "oharrak": {
                                        "description": "Oharrak edo komentarioak",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Kolorea ondo eguneratu da datubasean."
                    },
                    "404": {
                        "description": "Ez du kolorerik topatu ID horrekin."
                    }
                }
            },
            "post": {
                "tags": [
                    "Koloreak"
                ],
                "description": "Adierazitako informazioarekin kolore berri bat txertatzen du datubasean.",
                "operationId": "b9cb7f8e798104639eab3fa559a41ec9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "description": "Transakzioaren data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "id_bezeroa": {
                                        "description": "Bezeroaren ID-a",
                                        "type": "integer"
                                    },
                                    "id_produktua": {
                                        "description": "Produktuaren ID-a",
                                        "type": "integer"
                                    },
                                    "kantitatea": {
                                        "description": "Produktu kantitatea",
                                        "type": "integer"
                                    },
                                    "bolumena": {
                                        "description": "Produktuaren bolumena",
                                        "type": "integer"
                                    },
                                    "oharrak": {
                                        "description": "Oharrak edo komenlarioak",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produktua ondo txertatu da datubasean."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Koloreak"
                ],
                "description": "Adierazitako ID-a erabilita datubasean koloreen ezabatze logikoa egiten du.",
                "operationId": "bd72f1aef8c93e2136e945dcdf3b0420",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Elementuaren ID-a",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Koloreak ondo ezabatu dira datubasean."
                    },
                    "404": {
                        "description": "Ez dago kolorerik ID horrekin."
                    }
                }
            }
        },
        "/kolore/{id}": {
            "get": {
                "tags": [
                    "Koloreak"
                ],
                "description": "Adierazitako ID-aren arabera kolore bat bueltatzen du.",
                "operationId": "5b7c372455adcaf7f1bafc2f67ed9f77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kolorearen ID-a",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aukeratutako kolorea bueltatzen du."
                    },
                    "404": {
                        "description": "Ez du kolorerik topatu ID horrekin."
                    }
                }
            }
        },
...
    },
...
}